Client class has been added. Demo included in Client
