+++ src/action/Action.java
@@ -1,3 +1,7 @@
+package action;
+
+import img.Image;
+
 public interface Action {
     Image doAction(Image image);
 }
