+++ src/action/filter/Expose.java
@@ -1,5 +1,10 @@
-public class Shadow extends Filter {
-    Shadow(float ratio) {
+package action.filter;
+
+import img.Image;
+
+public class Expose extends Filter {
+
+    Expose(float ratio) {
         this.ratio = ratio;
     }
