+++ src/action/filter/Filter.java
@@ -1,3 +1,8 @@
+package action.filter;
+
+
+import action.Action;
+
 public abstract class Filter implements Action {
     protected float ratio;
 }
