+++ src/exporter/jpgExporter.java
@@ -1,6 +1,10 @@
+package exporter;
+
+import exporter.Exporter;
+
 import java.io.File;
 
-public class pngExporter implements Exporter {
+public class jpgExporter implements Exporter {
     @Override
     public File convertAndSave(String path) {
         return null;
