+++ src/exporter/tiffExporter.java
@@ -1,6 +1,10 @@
+package exporter;
+
+import exporter.Exporter;
+
 import java.io.File;
 
-public class jpgExporter implements Exporter {
+public class tiffExporter implements Exporter {
     @Override
     public File convertAndSave(String path) {
         return null;
