+++ src/img/ImgViewer.java
@@ -1,3 +1,5 @@
+package img;
+
 public class ImgViewer {
 
 }
