+++ src/img/Pixel.java
@@ -1,3 +1,5 @@
+package img;
+
 public class Pixel {
     char r, g, b;
     Pixel(char r, char g, char b) {
